Write POM to file and adjust to Darkestperu changes
Playing around with Clojars made me realize that I also need the pom.xml
external to the JAR. Add writing it out. Adjust to the changed API of
Darkestperu. (Actually, it was changed because of the POM thing, too. I
hope grenada-lib and Darkestperu got prettier together.)

Also make find-entities-clj use the ROOT argument and not something
hardcoded as before. – This I had overlooked when cleaning up after my
experiments.