Use Clojure 1.7 prn for sensible printing
Previously I had to remove objects that cannot be read before printing,
because pr printed them as #<…. Bump up the Clojure dependency to
1.7.0-RC1 (project.clj not yet under version control). In Clojure 1.7 pr
prints things as #object tagged literals if it doesn't know better. Thus
remove the manual (and very incomplete) removing of non-re-readable
objects. I'll just add the appropriate data-readers when the EDN files
have to be read again.