jar-from-project: Separate in and out coordinates
Previously you could only scan a project and produce an artifact with
the same coordinates as that project. I.e., if your Leiningen project
was [mygroup/myart "12.45.4"], you could only produce [mygroup/myart
"12.45.4" :classifier "metadata"]. This might be the most frequent
usecase, but separate coordinates for the input and output are necessary
at least occasionally.

My own tests are an example. I'm scanning [clj-grenada/grenada-lib
"0.1.0-SNAPSHOT"], but for my tests I don't want to push to the group
clj-grenada on Clojars. Now I can produce an artifact with coordinates
[org.clojars.rmoehn/grenada-lib "0.1.2"].