Correct the Clojure source
Once again I didn't make sure I know exactly how things work. It turns
out that resolve-dependencies, which I used in sources/clj.clj to
(download and) find the path to an artifact, gives me the paths to the
artifact and all its dependencies. Thus, when I ran
tns.find/find-namespaces on the result, it extracted a whole bunch of
namespace names that I didn't want. Use resolve-artifacts instead of
resolve dependencies, since this does what I originally wanted.

Now, resolve-artifacts is the right thing when we only want the path to
the artifact we want to extract metadata from. However, in order to load
the namespaces of this artifact inside the separate runtime, we also
have to load their dependencies. Thus, we need resolve-dependencies in
addition to resolve-artifacts, so that everyone is satisfied.

This leads to the next problem. We have to load all dependencies into
the separate runtime. This is no problem in cases where the thing we're
supposed to extract metadata from is specified as a Leiningen dependency
spec. In this case, Pomegranate will do all the legwork. But the
probably most frequent case is that people want to run the extraction on
their Leiningen project. Then we're facing a bunch of files and have no
clue about the dependencies. This is not difficult to solve, but I
hadn't thought of it before and will have to rearrange some code in
order to support this. For now, though, keep hacking away at other
things and only indicate what works and what doesn't.