Remove obsolete stuff, especially contracts
Remove the code that had to do with contracts, because it wasn't used
anymore and I don't know if it will be again. They were for specifying
how to customize the components or write your own. But I think I already
wrote somewhere else that that degree of customizability might result in
compatibility hell for me. With giving up this, the contracts became
unused.

While removing the contracts stuff I found some other things that
weren't used anymore. Remove those as well.