Fix coordinate parameters
I used to accept maps with coordinate parameters as {:group …, :name …,
:version …}. This is the format used by Leiningen. However, for reasons
of clarity and for not shadowing clojure.core bindings, I always
remapped "name" to "artifact" like this: {artifact :name :keys [group
version]}. This converting back and forth is silly and I don't approve
of the choice that was made in Leiningen anyway.

That's why in 00aa3cb I stopped accepting {:group …, :name …, :version
…} and required {:group …, :artifact …, :version …}. (Forgot to mention
it, though. :-() If you want to use Grenada with Leiningen now, you have
to convert the input up front and that's all. However, I didn't change
this everywhere and messed up the order in one place where I did change
it. Correct these things.