Adjust semantics to new constraints on data
grenada.things requires that the :extensions (and :cmeta) entries be
present in a Thing, even if they're just empty maps. Change strip-all to
empty the maps instead of removing the entries. – This slightly bloats
skeletons for external metadata, but it's small enough that I'd favour
standard-conformity.

dissoc-in removes empty maps and might therefore cause violation of the
abovementioned constraint. Change its last occurence to dissoc-in*.

Now that the stripped data still contain the :extensions (and :cmeta)
entries, we have to change the semantics of simple-merge. – The
simple-merge as it was before would complain about any occurence of
equal keys with different values in the maps to be merged. Change it to
make the merge of an empty thing with a non-empty thing be the non-empty
thing.

Also adapt the vector that determines the sort order of Things. –
Somehow fits with this commit.