Change validation semantics of tag-type?+
The predicate I was generating with deftag+ complected identification
and validation again. The result was that

 (A) we couldn't tell whether a tagged value was not of the right type
     for the purpose or was of the right type for the purpose, but
     misshapen;

 (B) if we actually found out it was misshapen, we couldn't know what
     was wrong, because we threw away the return value of
     schema.core/check. There was no other way than to throw it away,
     because in the end we had to return falsey and falsey can't carry
     any other information.

Hmm, of course you could return something that indicates misshapenness…
However, go the most direct way: understand a right-type-but-misshapen
value as an inconsistency and throw an exception in this case.

This might not yet be what we want.