Add a note to coords->path
I noticed for the first time the 1%2E6%2E0 path components and thought
it looked a bit silly and different from the Grimoire directories. I
found out that Grimoire only munges the Def coordinate and began to
change coords->path. But then I thought it doesn't make much more sense
that way. Therefore leave it that way and just add a clarifying note.