Fix argument to leiningen.pom/make-pom
I had changed my policy of coordinate naming in 00aa3cb (more
information in the commit message of b335014). The decision was to not
comply with the Leiningen idiosyncrasy of calling artifactId :name. So
I'm now passing a map with an :artifact key to
grenada.core/jar-from-files. However, grenada.core/jar-from-files calls
back to Leiningen's make-pom, which does expect a :name key. I had
forgotten to provide for that. Fix it now.