grenada.core: Add a TODO about POM
I found that it's helpful to add a <description> to a Datadoc JAR's POM.
Luckily, leiningen.pom/make-pom supports this out of the box. No
additional code is necessary. However, this feature is undocumented in
jar-from-files and the argument name is not appropriate for this use
case. Add a TODO to remind me to improve that.