Add above/below-incl? and aspect support
We had above-incl? and below-incl? before, but I removed them, because
they didn't fit with the new model. Now they're required again for the
Bar type definitions. Add them, adapted to the new model. Unfortunately,
they can't be as simple as before for two reasons:

 - Given a Thing T, its Main Aspect can't be determined directly. We
   have to look which Aspect of T's is on list of Main Aspects.

 - The Aspect prerequisites predicate only gets a set of Aspects as its
   argument and not a whole Thing. Therefore one can't use the usual
   functions for Things. That's why we need resolve-main-aspect to
   differentiate between the two cases.

   The reason why the predicate gets only a set of Aspects is to prevent
   people from checking other properties of a Thing. Since this is not
   so important, I would sacrifice it in favour of less tangled code in
   case we need more functions for dealing with Aspects that are also
   defined for Things.

Also correct the use of below-incl? in grenada.bars.