Improve asserts in a few spots
Include the arguments in the diagnostics strings of asserts that failed
during development. This helps identify culprits. Do this only in the
places where asserts failed during development. – Right now I don't have
the time to repair this everywhere, but it will be done in a visible
future.

Also use schema.macros/assert! instead of clojure.core/assert, since it
throws RuntimeExceptions, which are more appropriate to argument
checking than AssertionErrors. I know, using assert for checking
arguments is abusing it and assert! is only a small improvement.
However, I don't know of a more correct way that is not ugly. (Ooh,
well, there's always dependent types.)