Deprecate Grey
I hadn't searched well enough for a condition/restart system in Clojure.
There is Hara Event, which I haven't looked at much, but which should
work. Deprecate Grey. – Unless it's really horrible, there is no need
for an ultra-primitive reinvention of the wheel. – Add a TODO that we
should convert.